ml/requirements.txt
psycopg2-binary
pandas
numpy
bottle
